SELECT and EXAMINE processors now return stream of response. Removed deprecated response and encoder. No longer send additional unnecessary message when no messages are seen.
git-svn-id: https://svn.apache.org/repos/asf/james/server/trunk@642700 13f79535-47bb-0310-9956-ffa450edef68

+++ imap-codec-library/src/main/java/org/apache/james/imapserver/codec/encode/imap4rev1/StatusResponseEncoder.java
@@ -50,7 +50,7 @@ public class StatusResponseEncoder extends AbstractChainedImapEncoder {
         final HumanReadableTextKey textKey = response.getTextKey();
         final String text = asString(textKey);
         final Collection parameters;
-        final int number;
+        final long number;
         if (responseCode == null) {
             parameters = null;
             number = 0;
